+++ src/main/java/net/spell_power/SpellPowerMod.java
@@ -1,14 +1,24 @@
 package net.spell_power;
 
 import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.item.v1.EnchantmentEvents;
+import net.fabricmc.fabric.api.util.TriState;
+import net.minecraft.component.ComponentType;
+import net.minecraft.component.DataComponentTypes;
+import net.minecraft.component.EnchantmentEffectComponentTypes;
+import net.minecraft.component.type.AttributeModifiersComponent;
 import net.minecraft.entity.attribute.EntityAttributeModifier;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
 import net.minecraft.util.Identifier;
 import net.spell_power.api.*;
 import net.spell_power.config.AttributesConfig;
+import net.spell_power.internals.AttributeUtil;
 import net.spell_power.internals.CrossFunctionalAttributes;
 import net.tinyconfig.ConfigManager;
+
+import java.util.Set;
+import java.util.stream.Collectors;
 
 public class SpellPowerMod implements ModInitializer {
     public static final String ID = "spell_power";
@@ -44,6 +54,29 @@ public class SpellPowerMod implements ModInitializer {
                 CrossFunctionalAttributes.power(school.attributeEntry, genericSpellSchool.attributeEntry);
             }
         }
+
+        EnchantmentEvents.ALLOW_ENCHANTING.register((enchantment, target, enchantingContext) -> {
+            if (enchantment.isIn(SpellPowerTags.Enchantments.SCHOOL_FILTERED))  {
+                System.out.println("Spell Power - School Filtering check: " + enchantment);
+                var enchantmentAttributes = enchantment.value().effects().get(EnchantmentEffectComponentTypes.ATTRIBUTES);
+
+
+                if (enchantmentAttributes != null && !enchantmentAttributes.isEmpty()) {
+                    var itemAttributes = target.getComponents().get(DataComponentTypes.ATTRIBUTE_MODIFIERS);
+                    if (itemAttributes.modifiers().isEmpty()) {
+                        itemAttributes = target.getItem().getAttributeModifiers();
+                    }
+                    System.out.println("Spell Power - School Filtering | enchantmentAttributes: " + enchantmentAttributes + " | itemAttributes: " + itemAttributes);
+
+                    var intersect = AttributeUtil.attributesIntersect(enchantmentAttributes, itemAttributes);
+                    System.out.println("Spell Power - Intersect: " + intersect + " | " + enchantmentAttributes + " | " + itemAttributes);
+                    if (!intersect) {
+                        return TriState.FALSE;
+                    }
+                }
+            }
+            return TriState.DEFAULT;
+        });
     }
 
     /**