+++ src/main/java/net/spell_power/api/SpellPowerTags.java
@@ -1,12 +1,17 @@
 package net.spell_power.api;
 
+import net.minecraft.enchantment.Enchantment;
+import net.minecraft.entity.damage.DamageType;
 import net.minecraft.registry.RegistryKeys;
 import net.minecraft.registry.tag.TagKey;
 import net.minecraft.util.Identifier;
 import net.spell_power.SpellPowerMod;
 
 public class SpellPowerTags {
-    public static class DamageType {
-        public static final TagKey<net.minecraft.entity.damage.DamageType> ALL = TagKey.of(RegistryKeys.DAMAGE_TYPE, Identifier.of(SpellPowerMod.ID, "all"));
+    public static class DamageTypes {
+        public static final TagKey<DamageType> ALL = TagKey.of(RegistryKeys.DAMAGE_TYPE, Identifier.of(SpellPowerMod.ID, "all"));
+    }
+    public static class Enchantments {
+        public static final TagKey<Enchantment> SCHOOL_FILTERED = TagKey.of(RegistryKeys.ENCHANTMENT, Identifier.of(SpellPowerMod.ID, "school_filtered"));
     }
 }